fix: correct hallucinated NuGet versions & migrate LanceDB 0.x→2.x API
NuGet Version Fixes (5 packages corrected):
- Microsoft.WindowsAppSDK: 1.7.250405001 (hallucinated) → 1.7.260224002
- Microsoft.Windows.CsWinRT: 2.2.1 (hallucinated) → 2.2.0
- CommunityToolkit.Mvvm: 8.4.0 (outdated) → 8.4.2
- CommunityToolkit.WinUI.UI.Controls: 8.2.0 (hallucinated) → 7.1.2
- LanceDB: 0.14.0 (hallucinated) → 2.3.0
- Removed nuget.config (dotnet-public feed no longer needed)

LanceDB API Migration (LanceVectorDbService.cs full rewrite):
- Connection: new LanceDBConnection(path) → new Connection() + Connect(path)
- Data format: POCO rows → Apache Arrow RecordBatch with FixedSizeListType vectors
- Search: .Search(vector).ToResultSet() → .Query().NearestTo(vector).ToList()
- All ops now natively async (removed Task.Run wrappers)
- Schema: id, source_text, target_text, source_language, target_language,
  domain, quality_score, created_at_unix, vector (FixedSizeList[float32])

Build-fixes for Core project (0 errors, 5 warnings):
- Added Microsoft.Extensions.Logging.Abstractions + Microsoft.Extensions.Http to Core csproj
- Fixed ILocalInferenceService missing using for LlmState enum
- Fixed 10 interface implementation mismatches (CancellationToken params)
- Fixed OnnxEmbeddingService: flattened 2D arrays, DenseTensor API, NamedOnnxValue
- Fixed TmImportService: CsvHelper 33.x HeaderRecord API change
- Fixed RagPipelineService: long→int cast for dictionary initializer
- Fixed OnnxLlmInferenceService: dynamic dispatch on LogWarning

Verified: dotnet build src/RDAT.Copilot.Core succeeds with 0 errors.

+++ src/RDAT.Copilot.Core/Services/RagPipelineService.cs
@@ -301,7 +301,7 @@ public sealed class RagPipelineService : IRagPipelineService, IDisposable
             DomainBreakdown: new Dictionary<string, int>(),
             LanguagePairs: new Dictionary<string, int>
             {
-                { "en→ar", count },
+                { "en→ar", (int)count },
                 { "ar→en", 0 }
             },
             DbSizeMb: dbSizeMb
